` add maintence page and session time out `

+++ src/components/Logout.jsx
@@ -7,12 +7,30 @@ import {
 } from "@material-tailwind/react";
 import React from "react";
 import { useNavigate } from "react-router-dom";
+import BASE_URL from "../base/BaseUrl";
+import axios from "axios";
+import { toast } from "react-toastify";
 
 const Logout = ({ open, handleOpen }) => {
   const navigate = useNavigate();
-  const handleLogout = () => {
-    localStorage.clear();
-    navigate("/");
+
+  const handleLogout = async () => {
+    try {
+      const res = await axios.post(`${BASE_URL}/api/panel-logout`,{}, {
+        headers: {
+          Authorization: `Bearer ${localStorage.getItem("token")}`,
+        },
+      });
+      if (res.data.code === 200) {
+        toast.success(res.data.msg);
+        localStorage.clear();
+        navigate("/");
+      } else {
+        toast.error(res.data.msg);
+      }
+    } catch (error) {
+      console.error("Logout failed:", error);
+    }
   };
 
   return (
